+++ src/main/java/insane96mcp/insanesurvivaloverhaul/mixin/accessor/BlockStateBaseAccessor.java
@@ -2,10 +2,15 @@ package insane96mcp.insanesurvivaloverhaul.mixin.accessor;
 
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(BlockBehaviour.BlockStateBase.class)
 public interface BlockStateBaseAccessor {
     @Accessor
     float getDestroySpeed();
+
+    @Accessor
+    @Mutable
+    void setDestroySpeed(float destroySpeed);
 }